requirements.txt
Pillow==3.3.1
PyYAML==3.11
beautifulsoup4==4.5.1
cssselect==0.9.2
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.1
jieba3k==0.35.1
lxml==3.6.4
newspaper3k==0.1.7
nltk==3.2.1
python-dateutil==2.5.3
requests==2.11.1
requests-file==1.4
six==1.10.0
tldextract==2.0.1
